types/btctypes: moved address, tx, utxo into btctypes
