+++ types/btctypes/address.go
@@ -1,15 +1,17 @@
-package btcaddress
+package btctypes
 
 import (
 	"crypto/ecdsa"
+	"crypto/sha256"
 	"fmt"
 
 	"github.com/btcsuite/btcd/btcec"
+	"github.com/btcsuite/btcd/chaincfg"
 	"github.com/btcsuite/btcd/txscript"
 	"github.com/btcsuite/btcutil"
+	"github.com/btcsuite/btcutil/base58"
 	"github.com/iqoption/zecutil"
 	"github.com/renproject/mercury/types"
-	"github.com/renproject/mercury/types/btctypes"
 )
 
 // Address is an interface type for any type of destination a transaction output may spend to. This includes pay-to-
@@ -19,22 +21,22 @@ type Address btcutil.Address
 
 type Recipient struct {
 	Address Address
-	Amount  btctypes.Amount
+	Amount  Amount
 }
 
 // NewRecipient creates a new recipient
-func NewRecipient(address Address, amount btctypes.Amount) Recipient {
+func NewRecipient(address Address, amount Amount) Recipient {
 	return Recipient{address, amount}
 }
 
 type Recipients []Recipient
 
 // SerializePublicKey serializes the public key to bytes.
-func SerializePublicKey(pubkey ecdsa.PublicKey, network btctypes.Network) []byte {
+func SerializePublicKey(pubkey ecdsa.PublicKey, network Network) []byte {
 	switch network {
-	case btctypes.BtcMainnet, btctypes.ZecMainnet:
+	case BtcMainnet, ZecMainnet:
 		return (*btcec.PublicKey)(&pubkey).SerializeCompressed()
-	case btctypes.BtcTestnet, btctypes.BtcLocalnet, btctypes.ZecTestnet, btctypes.ZecLocalnet:
+	case BtcTestnet, BtcLocalnet, ZecTestnet, ZecLocalnet:
 		return (*btcec.PublicKey)(&pubkey).SerializeUncompressed()
 	default:
 		panic(types.ErrUnknownNetwork)
@@ -42,7 +44,7 @@ func SerializePublicKey(pubkey ecdsa.PublicKey, network btctypes.Network) []byte
 }
 
 // AddressFromBase58 decodes the base58 encoded address to an `Address`.
-func AddressFromBase58(addr string, network btctypes.Network) (Address, error) {
+func AddressFromBase58(addr string, network Network) (Address, error) {
 	switch network.Chain() {
 	case types.Bitcoin:
 		return btcutil.DecodeAddress(addr, network.Params())
@@ -54,7 +56,7 @@ func AddressFromBase58(addr string, network btctypes.Network) (Address, error) {
 }
 
 // AddressFromPubKey gets the `Address` from a public key.
-func AddressFromPubKey(pubkey ecdsa.PublicKey, network btctypes.Network) (Address, error) {
+func AddressFromPubKey(pubkey ecdsa.PublicKey, network Network) (Address, error) {
 	switch network.Chain() {
 	case types.Bitcoin:
 		addr, err := btcutil.NewAddressPubKey(SerializePublicKey(pubkey, network), network.Params())
@@ -70,7 +72,7 @@ func AddressFromPubKey(pubkey ecdsa.PublicKey, network btctypes.Network) (Addres
 }
 
 // AddressFromScript gets the `Address` from a script.
-func AddressFromScript(script []byte, network btctypes.Network) (Address, error) {
+func AddressFromScript(script []byte, network Network) (Address, error) {
 	switch network.Chain() {
 	case types.Bitcoin:
 		addr, err := btcutil.NewAddressScriptHash(script, network.Params())
@@ -86,7 +88,7 @@ func AddressFromScript(script []byte, network btctypes.Network) (Address, error)
 }
 
 // PayToAddrScript gets the PayToAddrScript for an address on the given blockchain
-func PayToAddrScript(address Address, network btctypes.Network) ([]byte, error) {
+func PayToAddrScript(address Address, network Network) ([]byte, error) {
 	switch network.Chain() {
 	case types.Bitcoin:
 		return txscript.PayToAddrScript(address)
@@ -95,4 +97,44 @@ func PayToAddrScript(address Address, network btctypes.Network) ([]byte, error) 
 	default:
 		return nil, fmt.Errorf("unsupported blockchain: %v", network.Chain())
 	}
+}
+
+func zecAddressFromHash160(hash []byte, params *chaincfg.Params, isScript bool) (btcutil.Address, error) {
+	prefixes := map[string]map[string][]byte{
+		"mainnet": map[string][]byte{
+			"pubkey": []byte{0x1C, 0xB8},
+			"script": []byte{0x1C, 0xBD},
+		},
+		"testnet3": map[string][]byte{
+			"pubkey": []byte{0x1D, 0x25},
+			"script": []byte{0x1C, 0xBA},
+		},
+		"regtest": map[string][]byte{
+			"pubkey": []byte{0x1D, 0x25},
+			"script": []byte{0x1C, 0xBA},
+		},
+	}
+	if isScript {
+		return zecutil.DecodeAddress(encodeHash(hash[:], prefixes[params.Name]["script"]), params.Name)
+	}
+	return zecutil.DecodeAddress(encodeHash(hash[:], prefixes[params.Name]["pubkey"]), params.Name)
+}
+
+func encodeHash(addrHash, prefix []byte) string {
+	var (
+		body  = append(prefix, addrHash...)
+		chk   = addrChecksum(body)
+		cksum [4]byte
+	)
+	copy(cksum[:], chk[:4])
+	return base58.Encode(append(body, cksum[:]...))
+}
+
+func addrChecksum(input []byte) (cksum [4]byte) {
+	var (
+		h  = sha256.Sum256(input)
+		h2 = sha256.Sum256(h[:])
+	)
+	copy(cksum[:], h2[:4])
+	return
 }
