--- types/btctypes/btcaddress/btc.go
@@ -1,1 +1,0 @@
-package btcaddress
