--- types/btctypes/btctx/tx.go
@@ -1,13 +1,0 @@
-package btctx
-
-import (
-	"github.com/renproject/mercury/types"
-	"github.com/renproject/mercury/types/btctypes/btcaddress"
-	"github.com/renproject/mercury/types/btctypes/btcutxo"
-)
-
-type BtcTx interface {
-	types.Tx
-	UTXOs() btcutxo.UTXOs
-	OutputUTXO(address btcaddress.Address) btcutxo.UTXO
-}
